+++ cmd/kubeadm/app/componentconfigs/kubelet.go
@@ -17,10 +17,7 @@ limitations under the License.
 package componentconfigs
 
 import (
-	"os"
 	"path/filepath"
-	"runtime"
-	"strings"
 
 	"github.com/pkg/errors"
 	"k8s.io/apimachinery/pkg/util/version"
@@ -231,60 +228,6 @@ func (kc *kubeletConfig) Default(cfg *kubeadmapi.ClusterConfiguration, _ *kubead
 			}
 		}
 	}
-}
-
-// Mutate modifies absolute path fields in the KubeletConfiguration to be Windows compatible absolute paths.
-func (kc *kubeletConfig) Mutate() error {
-	// TODO: use build tags and move the Windows related logic to _windows.go files
-	// once the kubeadm code base is unit tested for Windows as part of CI - "GOOS=windows go test ...".
-	if runtime.GOOS != "windows" {
-		return nil
-	}
-
-	// When "kubeadm join" downloads the KubeletConfiguration from the cluster on Windows
-	// nodes, it would contain absolute paths that may lack drive letters, since the config
-	// could have been generated on a Linux control-plane node. On Windows the
-	// Golang path.IsAbs() function returns false unless the path contains a drive letter.
-	// This trips client-go and the kubelet, creating problems on Windows nodes.
-	// Fixing it in client-go or the kubelet is a breaking change to existing Windows
-	// users that rely on relative paths:
-	//   https://github.com/kubernetes/kubernetes/pull/77710#issuecomment-491989621
-	//
-	// Thus, a workaround here is to adapt the KubeletConfiguration paths for Windows.
-	// Note this is currently bound to KubeletConfiguration v1beta1.
-	klog.V(2).Infoln("[componentconfig] Adapting the paths in the KubeletConfiguration for Windows...")
-
-	// Get the drive from where the kubeadm binary was called.
-	exe, err := os.Executable()
-	if err != nil {
-		return errors.Wrap(err, "could not obtain information about the kubeadm executable")
-	}
-	drive := filepath.VolumeName(filepath.Dir(exe))
-	klog.V(2).Infof("[componentconfig] Assuming Windows drive %q", drive)
-
-	// Mutate the paths in the config.
-	mutatePathsOnWindows(&kc.config, drive)
-	return nil
-}
-
-func mutatePathsOnWindows(cfg *kubeletconfig.KubeletConfiguration, drive string) {
-	mutateStringField := func(name string, field *string) {
-		// path.IsAbs() is not reliable here in the Windows runtime, so check if the
-		// path starts with "/" instead. This means the path originated from a Unix node and
-		// is an absolute path.
-		if !strings.HasPrefix(*field, "/") {
-			return
-		}
-		// Prepend the drive letter to the path and update the field.
-		*field = filepath.Join(drive, *field)
-		klog.V(2).Infof("[componentconfig] kubelet/Windows: adapted path for field %q to %q", name, *field)
-	}
-
-	// Mutate the fields we care about.
-	klog.V(2).Infof("[componentconfig] kubelet/Windows: changing field \"resolverConfig\" to empty")
-	cfg.ResolverConfig = utilpointer.String("")
-	mutateStringField("staticPodPath", &cfg.StaticPodPath)
-	mutateStringField("authentication.x509.clientCAFile", &cfg.Authentication.X509.ClientCAFile)
 }
 
 // isServiceActive checks whether the given service exists and is running
