+++ cmd/kubeadm/app/componentconfigs/kubelet_test.go
@@ -298,67 +298,3 @@ func TestKubeletFromCluster(t *testing.T) {
 		return kubeletHandler.FromCluster(client, testClusterCfg(legacyKubeletConfigMap))
 	})
 }
-
-func TestMutatePathsOnWindows(t *testing.T) {
-	const drive = "C:"
-	var fooResolverConfig string = "/foo/resolver"
-
-	tests := []struct {
-		name     string
-		cfg      *kubeletconfig.KubeletConfiguration
-		expected *kubeletconfig.KubeletConfiguration
-	}{
-		{
-			name: "valid: all fields are absolute paths",
-			cfg: &kubeletconfig.KubeletConfiguration{
-				ResolverConfig: &fooResolverConfig,
-				StaticPodPath:  "/foo/staticpods",
-				Authentication: kubeletconfig.KubeletAuthentication{
-					X509: kubeletconfig.KubeletX509Authentication{
-						ClientCAFile: "/foo/ca.crt",
-					},
-				},
-			},
-			expected: &kubeletconfig.KubeletConfiguration{
-				ResolverConfig: utilpointer.String(""),
-				StaticPodPath:  filepath.Join(drive, "/foo/staticpods"),
-				Authentication: kubeletconfig.KubeletAuthentication{
-					X509: kubeletconfig.KubeletX509Authentication{
-						ClientCAFile: filepath.Join(drive, "/foo/ca.crt"),
-					},
-				},
-			},
-		},
-		{
-			name: "valid: some fields are not absolute paths",
-			cfg: &kubeletconfig.KubeletConfiguration{
-				ResolverConfig: &fooResolverConfig,
-				StaticPodPath:  "./foo/staticpods", // not an absolute Unix path
-				Authentication: kubeletconfig.KubeletAuthentication{
-					X509: kubeletconfig.KubeletX509Authentication{
-						ClientCAFile: "/foo/ca.crt",
-					},
-				},
-			},
-			expected: &kubeletconfig.KubeletConfiguration{
-				ResolverConfig: utilpointer.String(""),
-				StaticPodPath:  "./foo/staticpods",
-				Authentication: kubeletconfig.KubeletAuthentication{
-					X509: kubeletconfig.KubeletX509Authentication{
-						ClientCAFile: filepath.Join(drive, "/foo/ca.crt"),
-					},
-				},
-			},
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			mutatePathsOnWindows(test.cfg, drive)
-			if !reflect.DeepEqual(test.cfg, test.expected) {
-				t.Errorf("Missmatch between expected and got:\nExpected:\n%+v\n---\nGot:\n%+v",
-					test.expected, test.cfg)
-			}
-		})
-	}
-}
